# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Señaventura",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/palabra/{id_nivel}": {
      "get": {
        "tags": [
          "Palabra"
        ],
        "description": "Endpoint para obtener las palabras de un nivel",
        "parameters": [
          {
            "name": "id_nivel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id del nivel"
          }
        ],
        "responses": {
          "200": {
            "description": "Palabras encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "palabra": {
                        "type": "string"
                      },
                      "id_video_segna": {
                        "type": "string"
                      },
                      "url_icono": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/palabra/": {
      "get": {
        "tags": [
          "Palabra"
        ],
        "description": "Endpoint para obtener todas las palabras",
        "responses": {
          "200": {
            "description": "Palabras encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "palabra": {
                        "type": "string"
                      },
                      "id_video_segna": {
                        "type": "string"
                      },
                      "url_icono": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Palabra"
        ],
        "description": "Endpoint para crear una palabra",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Error message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "palabra": {
                    "example": "any"
                  },
                  "id_video_segna": {
                    "example": "any"
                  },
                  "id_nivel": {
                    "example": "any"
                  },
                  "url_icono": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/palabra/{id_palabra}": {
      "delete": {
        "tags": [
          "Palabra"
        ],
        "description": "Endpoint para eliminar una palabra",
        "parameters": [
          {
            "name": "id_palabra",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id de la palabra"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Error message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}